# backend/scout-analytics-api-flask/openapi.yaml
openapi: 3.0.3
info:
  title: Scout Analytics API
  description: |
    Retail Intelligence API for Scout Analytics MVP
    
    This API provides comprehensive retail analytics capabilities including:
    - Transaction analysis and trends
    - Product mix and category insights
    - Consumer demographics and behavior
    - Brand substitution patterns
    - Store performance metrics
    
    Built with Flask and designed for Philippine retail market analysis.
  version: 3.0.0
  contact:
    name: Scout Analytics Team
    email: support@scout-analytics.com
  license:
    name: Proprietary
servers:
  - url: https://g8h3ilc786zz.manus.space/api
    description: Production server
  - url: http://localhost:8000/api
    description: Development server
paths:
...
  /analytics/overview:
    get:
      summary: Dashboard Overview Analytics
      description: Dashboard KPIs and summary metrics for overview page
      operationId: getOverviewAnalytics
      tags:
        - Analytics
      responses:
        '200':
          description: Overview analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      total_revenue:
                        type: number
                        format: float
                        example: 2847392.50
                        description: Total revenue in Philippine Pesos
                      transaction_count:
                        type: integer
                        example: 15000
                        description: Total number of transactions
                      avg_order_value:
                        type: number
                        format: float
                        example: 189.83
                        description: Average order value in Philippine Pesos
                      customer_count:
                        type: integer
                        example: 12750
                        description: Total unique customers
                      revenue_trend:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date
                            value:
                              type: number
                              format: float
                        description: 6-month revenue trend data
                      top_products:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                            revenue:
                              type: number
                              format: float
                            rank:
                              type: integer
                        description: Top 5 products by revenue
                      ai_insights:
                        type: array
                        items:
                          $ref: '#/components/schemas/AIInsight'
                        description: AI-generated business insights
                  metadata:
                    $ref: '#/components/schemas/ResponseMetadata'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /analytics/trends:
    get:
      summary: Transaction Trends Analytics
      description: Temporal analysis and regional distribution
      operationId: getTrendsAnalytics
      tags:
        - Analytics
      responses:
        '200':
          description: Trends analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      hourly_volume:
                        type: array
                        items:
                          type: object
                          properties:
                            hour:
                              type: integer
                              minimum: 0
                              maximum: 23
                            transaction_count:
                              type: integer
                        description: Transaction volume by hour of day
                      regional_distribution:
                        type: array
                        items:
                          type: object
                          properties:
                            region:
                              type: string
                            count:
                              type: integer
                            percentage:
                              type: number
                              format: float
                        description: Transaction distribution by region
                      peak_hours:
                        type: object
                        properties:
                          morning_peak:
                            type: object
                            properties:
                              hours:
                                type: string
                                example: "6-9 AM"
                              transaction_count:
                                type: integer
                                example: 2847
                          lunch_peak:
                            type: object
                            properties:
                              hours:
                                type: string
                                example: "11 AM-2 PM"
                              transaction_count:
                                type: integer
                                example: 3456
                          evening_peak:
                            type: object
                            properties:
                              hours:
                                type: string
                                example: "6-8 PM"
                              transaction_count:
                                type: integer
                                example: 4123
                      weekly_patterns:
                        type: array
                        items:
                          type: object
                          properties:
                            day:
                              type: string
                              enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
                            avg_transactions:
                              type: number
                              format: float
                  metadata:
                    $ref: '#/components/schemas/ResponseMetadata'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /analytics/products:
    get:
      summary: Product Mix Analytics
      description: Product mix and category analysis
      operationId: getProductAnalytics
      tags:
        - Analytics
      responses:
        '200':
          description: Product analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      category_distribution:
                        type: array
                        items:
                          type: object
                          properties:
                            category:
                              type: string
                            percentage:
                              type: number
                              format: float
                            revenue:
                              type: number
                              format: float
                        description: Product category distribution
                      brand_performance:
                        type: array
                        items:
                          type: object
                          properties:
                            brand:
                              type: string
                            revenue:
                              type: number
                              format: float
                            rank:
                              type: integer
                        description: Brand performance ranking
                      top_products:
                        type: array
                        items:
                          type: object
                          properties:
                            product:
                              type: string
                            revenue:
                              type: number
                              format: float
                            transactions:
                              type: integer
                        description: Top performing products
                      substitution_summary:
                        type: array
                        items:
                          type: object
                          properties:
                            from_brand:
                              type: string
                            to_brand:
                              type: string
                            count:
                              type: integer
                        description: Brand substitution summary
                  metadata:
                    $ref: '#/components/schemas/ResponseMetadata'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    ResponseMetadata:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of response
        request_id:
          type: string
          description: Unique identifier for request tracking
        processing_time_ms:
          type: integer
          description: Processing time in milliseconds
      required:
        - timestamp
        - request_id
        - processing_time_ms
...
    AIInsight:
      type: object
      properties:
        category:
          type: string
          enum: [Operations, Marketing, "Customer Insights", Inventory, "Sales Strategy"]
          description: Insight category
        message:
          type: string
          description: Human-readable insight message
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Confidence score as percentage
        actions:
          type: array
          items:
            type: string
          description: Recommended action items
      required:
        - category
        - message
        - confidence
        - actions
...
    Error:
      type: object
      properties:
        status:
          type: string
          example: "error"
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: string
              description: Additional error context
          required:
            - code
            - message
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
      required:
        - status
        - error
        - metadata
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: "error"
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
              details: "Please try again later or contact support"
...